#4  Refactor - remove unused entities and connected with them classes
