+++ src/main/java/pl/gamefactory/gameOfCard/controller/DeckController.java
@@ -1,9 +1,8 @@
-package pl.gamefactory.gameOfCard.resources;
+package pl.gamefactory.gameOfCard.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.web.bind.annotation.*;
 import pl.gamefactory.gameOfCard.entities.Deck;
 import pl.gamefactory.gameOfCard.entities.enums.Cards;
@@ -12,7 +11,6 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.Arrays;
-import java.util.List;
 
 @RestController
 @RequestMapping(value = "/api")
